requirements.txt
fastapi
uvicorn[standard]
pydantic
pdfplumber
python-multipart
python-dotenv
pinecone>=3.0.0
google-genai
anthropic
